+++ example/app-1/index.js
@@ -4,10 +4,13 @@ var app = express();
 var packageJson = require("./package.json");
 var name = packageJson.name;
 
+var id = Math.floor(Math.random() * 10000);
+
 var count = 0;
 app.get("*", function(req, res) {
   count += 1;
-  res.send("<code>" + name + "@" + packageJson.version + ": Loaded: " + count + "</code>");
+  const html = "<code>" + name + "@" + packageJson.version + ": Loaded: " + count + " | id: " + id + "</code><hr/>";
+  res.send(html);
 });