+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [Unreleased] - YYYY-MM-DD
 #### Added
 #### Changed
@@ -9,6 +10,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Removed
 #### Fixed
 #### Security
+
+
+## [1.2.0] - 2015-12-16
+#### Added
+- Added secret `tokens` for locking down the API (specified within `manifest.yml`).
+- API Action: Update
+- API Action: Restart
 
 
 ## [1.1.0] - 2015-12-14